axis-ptz/requirements.txt
coloredlogs==14.0
numpy-quaternion==2022.4.2
paho-mqtt==1.5.0
pandas==1.5.2
pytest==7.2.0
python-dateutil==2.8.2
sensecam-control==0.1.0
schedule==1.1.0
